Move ZenTitleBlock & ZenNavigationTabs to Header
