+++ draft-packages/header/KaizenDraft/Header/index.ts
@@ -1,0 +1,1 @@
+export { default as TitleBlock } from "./TitleBlock"
