--- draft-packages/zen-title-block/KaizenDraft/ZenTitleBlock/index.ts
@@ -1,1 +1,0 @@
-export { default as ZenTitleBlock } from "./ZenTitleBlock"
