--- draft-packages/zen-title-block/index.ts
@@ -1,1 +1,0 @@
-export * from "./KaizenDraft/ZenTitleBlock"
